.tools/pyproject.toml
[project]
name = 'icons-tools'
version = '0'
dependencies = [
    "pillow>=12.0.0",
]
